move BytesBacked to protocols file to be in sync with Als PR
