+++ algopy_testing/src/algopy_testing/primitives/__init__.py
@@ -1,6 +1,6 @@
 from algopy_testing.primitives.biguint import BigUInt
 from algopy_testing.primitives.bytes import Bytes
-from algopy_testing.primitives.string import BytesBacked, String
+from algopy_testing.primitives.string import String
 from algopy_testing.primitives.uint64 import UInt64
 
-__all__ = ["UInt64", "Bytes", "BigUInt", "String", "BytesBacked"]
+__all__ = ["UInt64", "Bytes", "BigUInt", "String"]
