+++ algopy_testing/src/algopy_testing/primitives/biguint.py
@@ -4,8 +4,8 @@ import functools
 
 from algopy_testing.constants import UINT64_BYTES_LENGTH
 from algopy_testing.primitives.bytes import Bytes
-from algopy_testing.primitives.string import BytesBacked
 from algopy_testing.primitives.uint64 import UInt64
+from algopy_testing.protocols import BytesBacked
 from algopy_testing.utils import as_bytes, as_int, as_int512, int_to_bytes
 
 # TypeError, ValueError are used for operations that are compile time errors
