+++ algopy_testing/src/algopy_testing/primitives/string.py
@@ -1,21 +1,8 @@
 from __future__ import annotations
 
-import typing
-
 from algopy_testing.primitives.bytes import Bytes
+from algopy_testing.protocols import BytesBacked
 from algopy_testing.utils import as_bytes, as_string
-
-
-class BytesBacked(typing.Protocol):
-    """Represents a type that is a single bytes value"""
-
-    @classmethod
-    def from_bytes(cls, value: Bytes | bytes, /) -> typing.Self:
-        """Construct an instance from the underlying bytes (no validation)"""
-
-    @property
-    def bytes(self) -> Bytes:
-        """Get the underlying Bytes"""
 
 
 class String(BytesBacked):
